+++ Assembler/src/transforms/register_reuse.rs
@@ -221,16 +221,13 @@ fn clean(map: &mut BTreeMap<u32, NonZeroU32>) {
                 return true;
             }
             // Is the register part of the current registers vectorized elements?
-            if r < cur_r + u32::from(cur_v.get()) {
+            if r < cur_r + cur_v.get() {
                 // Is the register *also* vectorized and goes beyond the current vector register?
-                let end = r + u32::from(v.get());
-                let cur_end = cur_r + u32::from(cur_v.get());
+                let end = r + v.get();
+                let cur_end = cur_r + cur_v.get();
                 if end > cur_end {
                     // expand register
-                    cur_v = NonZeroU32::new(
-                        u32::from(cur_v.get()) + end - cur_end
-                    )
-                    .unwrap();
+                    cur_v = NonZeroU32::new(cur_v.get() + end - cur_end).unwrap();
                 }
                 // do not keep this around, it's covered by the current register
                 false
@@ -278,7 +275,7 @@ impl super::Pass for Pass {
                         // FIXME: this is a bit inefficient (O(n^2)), maybe return a element register -> base register map from `clean`
                         // instead of computing `base_map` on the fly here
                         for (&base_r, base_v) in block_seen {
-                            for r in base_r..base_r + u32::from(base_v.get()) {
+                            for r in base_r..base_r + base_v.get() {
                                 base_map.insert(r, base_r);
                             }
                             block_seen_bases.insert(base_r);
@@ -346,7 +343,7 @@ impl super::Pass for Pass {
             .map(|(cb, seen)| -> u32 {
                 cb.iter()
                     // reserve multiple registers for all vectorized registers
-                    .map(|r| u32::from(seen[r].get()))
+                    .map(|r| seen[r].get())
                     .sum()
             });
 
@@ -386,7 +383,7 @@ impl super::Pass for Pass {
                         let mut allocate = |size: NonZeroU32, dealloc: Vec<(u32, NonZeroU32)>| {
                             trace!(?free, ?size, "allocate");
                             for (dealloc, size) in dealloc {
-                                let size = u32::from(size.get());
+                                let size = size.get();
                                 trace!(dealloc);
                                 // See if we fit right behind a hole, before a hole or in the middle of two.
                                 let mut iter = free.iter();
@@ -415,7 +412,7 @@ impl super::Pass for Pass {
                                     }
                                 }
                             }
-                            let size = u32::from(size.get());
+                            let size = size.get();
                             trace!(?free, ?size, "allocate");
                             // Naive "smallest fit" algorithm
                             let smallest = free
@@ -519,13 +516,12 @@ impl super::Pass for Pass {
                     } else {
                         let new = i;
                         // Reserve space for the register
-                        i += u32::from(
-                            seen.get(&base)
-                                .unwrap_or_else(|| {
-                                    panic!("register {} not found in {:#?}", r, seen);
-                                })
-                                .get(),
-                        );
+                        i += seen
+                            .get(&base)
+                            .unwrap_or_else(|| {
+                                panic!("register {} not found in {:#?}", r, seen);
+                            })
+                            .get();
                         glob.insert(r, new);
                     }
                 }
